Be a little more strict about what we think is an URL.
If a multi-line HTML document was encountered before now with one line that
started with 'http' then we incorrectly tried to pass the entire document as
an URL to the htmldoc command.

We could be a lot more strict - checking for newlines and other characters
that indicate that it's not really an URL - but this makes my use-case pass
so it's good enough for the moment.